integration: move remaining single use assertions into tests
